Added code to download from gdrive

+++ e2Dist/e2Dist.py
@@ -14,23 +14,25 @@ rcParams['pdf.fonttype'] = 42
 
 import os
 from os import path
-import urllib.request
+import gdown
 
-_DATA = "/tmp/e2Dist_data/"
+_DATA = "tmp/e2Dist/data/"
+_RESULTS = "tmp/e2Dist/results/"
 
-def _download(url, filename):
-  if not path.exists(_DATA):
-    os.makedirs(_DATA)
-  out_file = path.join(_DATA, filename)
-  if not path.isfile(out_file):
-    urllib.request.urlretrieve(url, out_file)
+def _gdownload(share_link, filename):
+    file_id = share_link.split('/')[-2]
+    url = f"https://drive.google.com/uc?id={file_id}"
+    if not path.exists(_DATA):
+        os.makedirs(_DATA)
+    out_file = path.join(_DATA, filename)
+    if not path.isfile(out_file):
+        gdown.download(url, out_file, quiet="False")
+    return out_file
 
-url = "https://www.dropbox.com/s/8u3r2ny9o5vwvh1/trento_Pb_Pb_4000000_-x_6.4_-n_13.94_-k_1.19_-p_0.007_-d_0.7_-w_0.4.csv?dl=0"
+share_link = "https://drive.google.com/file/d/1K8nlHIcM5NDmjbRXxv5BDCxg95YgKX8h/view?usp=sharing"
 filename = "trento_Pb_Pb_4000000_-x_6.4_-n_13.94_-k_1.19_-p_0.007_-d_0.7_-w_0.4.csv"
 
-_download(url,filename)
-
-dataString = path.join(_DATA,filename)
+dataString = _gdownload(share_link,filename)
 data = pd.read_csv(dataString)
 nch = np.asarray(data.iloc[:,3])
 e2 = np.asarray(data.iloc[:,4])
@@ -44,7 +46,11 @@ fig, ax = plt.subplots()
 ax.plot(x,y,color='blue',linestyle='dashed',linewidth=2)
 ax.set_xlabel('Normalized e2')
 ax.set_title('Distribution for 0 to 5% centrality')
-figString = path.join(_DATA,"figure1.png")
-plt.savefig(figString,format='png')
+
+if not path.exists(_RESULTS):
+    os.makedirs(_RESULTS)
+figString = path.join(_RESULTS,"figure1.png")
+if not path.isfile(figString):
+    plt.savefig(figString,format='png')
 plt.show()
